requirements.txt
absl-py==0.11.0
anyio==2.1.0
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
black==20.8b1
bleach==3.3.0
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.4
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
configparser==5.0.1
contextlib2==0.6.0.post1
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
docker-pycreds==0.4.0
einops==0.2.0
entrypoints==0.3
filelock==3.0.12
gitdb==4.0.5
GitPython==3.1.13
gql==0.1.0
graphql-core==3.1.3
identify==1.5.13
idna==2.10
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-server==1.3.0
jupyterlab==3.0.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.2.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
ml-collections==0.1.0
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
nodeenv==1.5.0
notebook==6.2.0
numpy==1.18.5
nvidia-ml-py3==7.352.0
olefile==0.46
opt-einsum==0+untagged.61.gd905544.dirty
packaging==20.9
pandocfilters==1.4.2
parso==0.8.1
pathspec==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip==20.1.1
pre-commit==2.10.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.13.0
psutil==5.7.3
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.4
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.1
Send2Trash==1.5.0
sentry-sdk==0.20.0
setuptools==49.2.0.post20200714
shortuuid==1.0.1
six==1.15.0
smmap==3.0.5
sniffio==1.2.0
subprocess32==3.5.4
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
torch==1.6.0
torchvision==0.7.0
tornado==6.1
tqdm==4.48.2
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.2
wandb==0.10.18
watchdog==0.10.4
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.34.2
zipp==3.4.0
